Bump up package to v2
- Rename package iampolicy to policy
- Move iam/policy to policy/
