+++ certs/certs_test.go
@@ -26,7 +26,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/minio/pkg/certs"
+	"github.com/minio/pkg/v2/certs"
 )
 
 func updateCerts(crt, key string) {
