+++ policy/action.go
@@ -15,11 +15,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package iampolicy
+package policy
 
 import (
-	"github.com/minio/pkg/iam/policy/condition"
-	"github.com/minio/pkg/wildcard"
+	"github.com/minio/pkg/v2/policy/condition"
+	"github.com/minio/pkg/v2/wildcard"
 )
 
 // Action - policy action.
