+++ policy/condition/boolfunc.go
