+++ policy/condition/boolfunc_test.go
