+++ policy/condition/func.go
