+++ policy/condition/func_test.go
