+++ policy/condition/ipaddrfunc_test.go
