+++ policy/condition/nullfunc.go
