+++ policy/condition/nullfunc_test.go
