+++ policy/condition/numericfunc_test.go
