+++ policy/condition/stringfunc_test.go
