+++ policy/condition/value_test.go
