+++ policy/effect_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2015-2023 MinIO, Inc.
+// Copyright (c) 2015-2021 MinIO, Inc.
 //
 // This file is part of MinIO Object Storage stack
 //
@@ -15,27 +15,49 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package iampolicy
+package policy
 
 import (
 	"testing"
 )
 
-func TestIDIsValid(t *testing.T) {
+func TestEffectIsAllowed(t *testing.T) {
 	testCases := []struct {
-		id             ID
+		effect         Effect
+		check          bool
 		expectedResult bool
 	}{
-		{ID("DenyEncryptionSt1"), true},
-		{ID(""), true},
-		{ID("aa\xe2"), false},
+		{Allow, false, false},
+		{Allow, true, true},
+		{Deny, false, true},
+		{Deny, true, false},
 	}
 
 	for i, testCase := range testCases {
-		result := testCase.id.IsValid()
+		result := testCase.effect.IsAllowed(testCase.check)
 
 		if result != testCase.expectedResult {
-			t.Errorf("case %v: result: expected: %v, got: %v\n", i+1, testCase.expectedResult, result)
+			t.Fatalf("case %v: expected: %v, got: %v\n", i+1, testCase.expectedResult, result)
+		}
+	}
+}
+
+func TestEffectIsValid(t *testing.T) {
+	testCases := []struct {
+		effect         Effect
+		expectedResult bool
+	}{
+		{Allow, true},
+		{Deny, true},
+		{Effect(""), false},
+		{Effect("foo"), false},
+	}
+
+	for i, testCase := range testCases {
+		result := testCase.effect.IsValid()
+
+		if result != testCase.expectedResult {
+			t.Fatalf("case %v: expected: %v, got: %v\n", i+1, testCase.expectedResult, result)
 		}
 	}
 }
