+++ policy/resource.go
@@ -15,15 +15,15 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package iampolicy
+package policy
 
 import (
 	"encoding/json"
 	"path"
 	"strings"
 
-	"github.com/minio/pkg/iam/policy/condition"
-	"github.com/minio/pkg/wildcard"
+	"github.com/minio/pkg/v2/policy/condition"
+	"github.com/minio/pkg/v2/wildcard"
 )
 
 // ResourceARNPrefix - resource ARN prefix as per AWS S3 specification.
