+++ policy/statement.go
@@ -15,32 +15,27 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package iampolicy
+package policy
 
 import (
 	"strings"
 
-	"github.com/minio/pkg/iam/policy/condition"
+	"github.com/minio/pkg/v2/policy/condition"
 )
 
-// BPStatement - policy statement.
-type BPStatement struct {
+// Statement - iam policy statement.
+type Statement struct {
 	SID        ID                  `json:"Sid,omitempty"`
 	Effect     Effect              `json:"Effect"`
-	Principal  Principal           `json:"Principal"`
 	Actions    ActionSet           `json:"Action"`
 	NotActions ActionSet           `json:"NotAction,omitempty"`
-	Resources  ResourceSet         `json:"Resource"`
+	Resources  ResourceSet         `json:"Resource,omitempty"`
 	Conditions condition.Functions `json:"Condition,omitempty"`
 }
 
 // IsAllowed - checks given policy args is allowed to continue the Rest API.
-func (statement BPStatement) IsAllowed(args BucketPolicyArgs) bool {
+func (statement Statement) IsAllowed(args Args) bool {
 	check := func() bool {
-		if !statement.Principal.Match(args.AccountName) {
-			return false
-		}
-
 		if (!statement.Actions.Match(args.Action) && !statement.Actions.IsEmpty()) ||
 			statement.NotActions.Match(args.Action) {
 			return false
@@ -53,9 +48,12 @@ func (statement BPStatement) IsAllowed(args BucketPolicyArgs) bool {
 			}
 
 			resource += args.ObjectName
+		} else {
+			resource += "/"
 		}
 
-		if !statement.Resources.Match(resource, args.ConditionValues) {
+		// For admin statements, resource match can be ignored.
+		if !statement.Resources.Match(resource, args.ConditionValues) && !statement.isAdmin() && !statement.isKMS() {
 			return false
 		}
 
@@ -65,33 +63,71 @@ func (statement BPStatement) IsAllowed(args BucketPolicyArgs) bool {
 	return statement.Effect.IsAllowed(check())
 }
 
+func (statement Statement) isAdmin() bool {
+	for action := range statement.Actions {
+		if AdminAction(action).IsValid() {
+			return true
+		}
+	}
+	return false
+}
+
+func (statement Statement) isKMS() bool {
+	for action := range statement.Actions {
+		if KMSAction(action).IsValid() {
+			return true
+		}
+	}
+	return false
+}
+
 // isValid - checks whether statement is valid or not.
-func (statement BPStatement) isValid() error {
+func (statement Statement) isValid() error {
 	if !statement.Effect.IsValid() {
 		return Errorf("invalid Effect %v", statement.Effect)
-	}
-
-	if !statement.Principal.IsValid() {
-		return Errorf("invalid Principal %v", statement.Principal)
 	}
 
 	if len(statement.Actions) == 0 && len(statement.NotActions) == 0 {
 		return Errorf("Action must not be empty")
 	}
 
+	if statement.isAdmin() {
+		if err := statement.Actions.ValidateAdmin(); err != nil {
+			return err
+		}
+		for action := range statement.Actions {
+			keys := statement.Conditions.Keys()
+			keyDiff := keys.Difference(adminActionConditionKeyMap[action])
+			if !keyDiff.IsEmpty() {
+				return Errorf("unsupported condition keys '%v' used for action '%v'", keyDiff, action)
+			}
+		}
+		return nil
+	}
+
+	if statement.isKMS() {
+		return statement.Actions.ValidateKMS()
+	}
+
+	if !statement.SID.IsValid() {
+		return Errorf("invalid SID %v", statement.SID)
+	}
+
 	if len(statement.Resources) == 0 {
 		return Errorf("Resource must not be empty")
 	}
 
+	if err := statement.Resources.Validate(); err != nil {
+		return err
+	}
+
+	if err := statement.Actions.Validate(); err != nil {
+		return err
+	}
+
 	for action := range statement.Actions {
-		if action.IsObjectAction() {
-			if !statement.Resources.ObjectResourceExists() {
-				return Errorf("unsupported Resource found %v for action %v", statement.Resources, action)
-			}
-		} else {
-			if !statement.Resources.BucketResourceExists() {
-				return Errorf("unsupported Resource found %v for action %v", statement.Resources, action)
-			}
+		if !statement.Resources.ObjectResourceExists() && !statement.Resources.BucketResourceExists() {
+			return Errorf("unsupported Resource found %v for action %v", statement.Resources, action)
 		}
 
 		keys := statement.Conditions.Keys()
@@ -105,20 +141,13 @@ func (statement BPStatement) isValid() error {
 }
 
 // Validate - validates Statement is for given bucket or not.
-func (statement BPStatement) Validate(bucketName string) error {
-	if err := statement.isValid(); err != nil {
-		return err
-	}
-
-	return statement.Resources.ValidateBucket(bucketName)
+func (statement Statement) Validate() error {
+	return statement.isValid()
 }
 
 // Equals checks if two statements are equal
-func (statement BPStatement) Equals(st BPStatement) bool {
+func (statement Statement) Equals(st Statement) bool {
 	if statement.Effect != st.Effect {
-		return false
-	}
-	if !statement.Principal.Equals(st.Principal) {
 		return false
 	}
 	if !statement.Actions.Equals(st.Actions) {
@@ -137,11 +166,10 @@ func (statement BPStatement) Equals(st BPStatement) bool {
 }
 
 // Clone clones Statement structure
-func (statement BPStatement) Clone() BPStatement {
-	return BPStatement{
+func (statement Statement) Clone() Statement {
+	return Statement{
 		SID:        statement.SID,
 		Effect:     statement.Effect,
-		Principal:  statement.Principal.Clone(),
 		Actions:    statement.Actions.Clone(),
 		NotActions: statement.NotActions.Clone(),
 		Resources:  statement.Resources.Clone(),
@@ -149,24 +177,22 @@ func (statement BPStatement) Clone() BPStatement {
 	}
 }
 
-// NewBPStatement - creates new statement.
-func NewBPStatement(sid ID, effect Effect, principal Principal, actionSet ActionSet, resourceSet ResourceSet, conditions condition.Functions) BPStatement {
-	return BPStatement{
+// NewStatement - creates new statement.
+func NewStatement(sid ID, effect Effect, actionSet ActionSet, resourceSet ResourceSet, conditions condition.Functions) Statement {
+	return Statement{
 		SID:        sid,
 		Effect:     effect,
-		Principal:  principal,
 		Actions:    actionSet,
 		Resources:  resourceSet,
 		Conditions: conditions,
 	}
 }
 
-// NewBPStatementWithNotAction - creates new statement with NotAction.
-func NewBPStatementWithNotAction(sid ID, effect Effect, principal Principal, notActions ActionSet, resources ResourceSet, conditions condition.Functions) BPStatement {
-	return BPStatement{
+// NewStatementWithNotAction - creates new statement with NotAction.
+func NewStatementWithNotAction(sid ID, effect Effect, notActions ActionSet, resources ResourceSet, conditions condition.Functions) Statement {
+	return Statement{
 		SID:        sid,
 		Effect:     effect,
-		Principal:  principal,
 		NotActions: notActions,
 		Resources:  resources,
 		Conditions: conditions,
